Move models/prompt.py -> tasks/agents
